+++ internal/interfaces/http/dtos/login_response.go
@@ -1,7 +1,7 @@
 package dtos
 
 type LoginResponse struct {
-	UserID       int64  `json:"user_id"`
+	UserID       int    `json:"user_id"`
 	Email        string `json:"email"`
 	Role         string `json:"role"`
 	AccessToken  string `json:"access_token"`
